feat: Remove Octo plugin from Neovim configuration
This commit removes the Octo plugin from the Neovim configuration. The
plugin was removed from the 'lazyvim.json' file and all the related
keybindings were commented out in the 'lua/plugins/git.lua' file. This
change is part of an effort to streamline the Neovim setup and remove
unused or unnecessary plugins.