feat(keymaps): add TitleComment command and keymap
This commit introduces a new command, TitleComment, which generates a
title comment with a given title. It also adds a new keymap for this
command. The title comment is a comment line that is decorated with
dashes and includes a title in the middle. This can be used to
visually separate different sections of the code.